Adding Init Script support, PreExecutionActions, and fixing bug
The bug prevents people from being able to specify the project dir
themselves using the annotation.

PreExecutionActions allow for people to cleanly extend nebula-test by adding
actions to to happen before builds. This can help when files need to be
written all the time, but the test sets up the data while running (can't
happen in the setup method)

And finally giving people the ability to add custom init scripts.