NEBULA-3606 Introduce TestKit DSL
The TestKit DSL provides a new abstraction over Gradle TestKit which is usable in Java, Groovy, and Kotlin, and can be used with any unit testing framework. It assists in setting up a sample project, with an API that makes the structure of the project easily understandable to the reader of the test. This implementation is in an MVP state, with more functionality planned to cover common use cases.

add the sample project "gradleTest" into the build as an includedBuild. remove samples using deprecated code.